Rename directory to pagination (not capitalized)
